+++ modules/tinymce/src/core/demo/ts/demo/StickyToolbarDemo.ts
@@ -72,7 +72,7 @@ export default () => {
       makeSidebar(ed, 'sidebar1', 'green', 200);
     },
     plugins: [
-      'autosave advlist autolink link image lists charmap preview hr anchor pagebreak',
+      'autosave advlist autolink link image lists charmap preview anchor pagebreak',
       'searchreplace wordcount visualblocks visualchars code fullscreen insertdatetime media nonbreaking',
       'save table directionality emoticons template paste importcss',
       'codesample help noneditable'