+++ modules/tinymce/src/core/test/ts/browser/init/InitEditorThemeFunctionInlineTest.ts
@@ -2,6 +2,7 @@ import { Assertions } from '@ephox/agar';
 import { describe, it } from '@ephox/bedrock-client';
 import { Insert, SelectorFind, SugarBody, SugarElement } from '@ephox/sugar';
 import { TinyAssertions, TinyDom, TinyHooks } from '@ephox/wrap-mcagar';
+import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
 
@@ -39,5 +40,10 @@ describe('browser.tinymce.core.init.InitEditorThemeFunctionInlineTest', () => {
     Assertions.assertDomEq('Should be expected editor body element', targetElement, TinyDom.body(editor));
     Assertions.assertDomEq('Should be expected editor target element', targetElement, TinyDom.targetElement(editor));
     Assertions.assertDomEq('Should be expected content area container', targetElement, TinyDom.contentAreaContainer(editor));
+  });
+
+  it('TINY-8348: no default content css loaded', () => {
+    const editor = hook.editor();
+    assert.lengthOf(editor.contentCSS, 0, 'Should not have loaded the default content css');
   });
 });