+++ modules/tinymce/src/plugins/link/main/ts/api/Options.ts
@@ -57,7 +57,7 @@ const register = (editor: Editor): void => {
 
   registerOption('link_default_protocol', {
     processor: 'string',
-    default: 'http'
+    default: 'https'
   });
 
   registerOption('link_target_list', {